Export everything from Table/index.tsx
Had to make all exports named.
