+++ src/components/Table/DrillDown/helpers/Pagination/pagination.tsx
@@ -2,7 +2,7 @@
 import { Dictionary } from '@onaio/utils';
 import React from 'react';
 import { OF, PAGE, ROWS_TO_DISPLAY } from '../../../../../configs/lang';
-import { RenderPaginationOptions } from '../../TableJSX';
+import { RenderPaginationOptions } from '../../components/TableJSX';
 import './pagination.css';
 
 /** interface describes props for Reveal Custom Pagination */
